[fix]useLayoutEffectをコンポーネントに移譲#3
